Basic Dialogflow agent.
This is the basic Dialogflow agent. we have a paid service which converts data to the required format. but we are looking to write code which will convert the given data to the require format. once data is uploaded we have all the code ready to get accuracy and recall.